add Alert to all controller and update CreateBillService (file chooser)

+++ src/main/java/th/ac/ku/app/controller/branch/BranchPageController.java
@@ -12,6 +12,7 @@ import javafx.stage.Modality;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 
+import javafx.stage.StageStyle;
 import th.ac.ku.app.controller.login.LoginController;
 import th.ac.ku.app.controller.orderInfo.OrderInfoController;
 import th.ac.ku.app.models.Cloth;
@@ -24,42 +25,63 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 public class BranchPageController {
 
-    @FXML private Button logoutBtn, logoutBtn1, logoutBtn2, logoutBtn3, logoutBtn4,
-                        infoBtn, infoBtn1, delBtn, clearOrderFieldBtn, addOrderBtn, showReportBtn,
-                        showClosedInfoBtn, changePasswordBtn, clearPasswordFieldBtn;
-    @FXML private TextField customerNameField, customerPhoneField, clothQuantityField;
-    @FXML private PasswordField currentPasswordField, newPasswordField, confirmPasswordField;
-    @FXML private TableView<OrderInfo> orderTable;
-    @FXML private TableColumn<OrderInfo, Integer> orderIdCol;
-    @FXML private TableColumn<OrderInfo, String > customerNameCol;
-    @FXML private TableColumn<OrderInfo, String > dateCol;
-    @FXML private TableColumn<OrderInfo, String > statusCol;
-    @FXML private TableView<OrderInfo> orderTable2;
-    @FXML private TableColumn<OrderInfo, Integer> orderIdCol2;
-    @FXML private TableColumn<OrderInfo, String > customerNameCol2;
-    @FXML private TableColumn<OrderInfo, String > dateCol2;
-    @FXML private TableColumn<OrderInfo, String > statusCol2;
-    @FXML private TableView<OrderInfo> closedOrderTable;
-    @FXML private TableColumn<Object, String > closedOrderIdCol;
-    @FXML private TableColumn<Object, String > closedCustomerNameCol;
-    @FXML private TableColumn<Object, Integer> quantityCol;
-    @FXML private Label branchNameLabel1,branchNameLabel2,branchNameLabel3,branchNameLabel4,branchNameLabel5;
+    @FXML
+    private Button logoutBtn, logoutBtn1, logoutBtn2, logoutBtn3, logoutBtn4,
+            infoBtn, infoBtn1, delBtn, clearOrderFieldBtn, addOrderBtn, showReportBtn,
+            showClosedInfoBtn, changePasswordBtn, clearPasswordFieldBtn;
+    @FXML
+    private TextField customerNameField, customerPhoneField, clothQuantityField;
+    @FXML
+    private PasswordField currentPasswordField, newPasswordField, confirmPasswordField;
+    @FXML
+    private TableView<OrderInfo> orderTable;
+    @FXML
+    private TableColumn<OrderInfo, Integer> orderIdCol;
+    @FXML
+    private TableColumn<OrderInfo, String> customerNameCol;
+    @FXML
+    private TableColumn<OrderInfo, String> dateCol;
+    @FXML
+    private TableColumn<OrderInfo, String> statusCol;
+    @FXML
+    private TableView<OrderInfo> orderTable2;
+    @FXML
+    private TableColumn<OrderInfo, Integer> orderIdCol2;
+    @FXML
+    private TableColumn<OrderInfo, String> customerNameCol2;
+    @FXML
+    private TableColumn<OrderInfo, String> dateCol2;
+    @FXML
+    private TableColumn<OrderInfo, String> statusCol2;
+    @FXML
+    private TableView<OrderInfo> closedOrderTable;
+    @FXML
+    private TableColumn<Object, String> closedOrderIdCol;
+    @FXML
+    private TableColumn<Object, String> closedCustomerNameCol;
+    @FXML
+    private TableColumn<Object, Integer> quantityCol;
+    @FXML
+    private Label branchNameLabel1, branchNameLabel2, branchNameLabel3, branchNameLabel4, branchNameLabel5;
 
     private AccountManager accountManager;
     private WashingOrderServiceAPI serviceAPI;
     private final DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd/MM/yyyy HH:mm:ss");
     private OrderInfo selectedOrder;
+    private Alert alert;
 
-    @FXML private void initialize(){
+    @FXML
+    private void initialize() {
         Platform.runLater(() -> {
             branchNameLabel1.setText(accountManager.getCurrentBranch().getName());
             branchNameLabel2.setText(accountManager.getCurrentBranch().getName());
             branchNameLabel3.setText(accountManager.getCurrentBranch().getName());
             branchNameLabel4.setText(accountManager.getCurrentBranch().getName());
-            branchNameLabel5.setText("hello, "+accountManager.getCurrentBranch().getName());
+            branchNameLabel5.setText("hello, " + accountManager.getCurrentBranch().getName());
             orderTable.setPlaceholder(new Label("Not have new order at this time"));
             orderTable2.setPlaceholder(new Label("Not have cleaned order at this time"));
             closedOrderTable.setPlaceholder(new Label("Not have closed order at this time"));
@@ -73,6 +95,11 @@ public class BranchPageController {
                     selectedOrderInfo(newValue);
                 }
             });
+            closedOrderTable.getSelectionModel().selectedItemProperty().addListener((observable, oldValue, newValue) -> {
+                if (newValue != null) {
+                    selectedOrderInfo(newValue);
+                }
+            });
             showOrderList();
             showCleanedOrderList();
             showClosedOrderList();
@@ -81,7 +108,7 @@ public class BranchPageController {
 
 //Preparing Page-------------------------------------------------------------------------------------------------------------
 
-    public void showOrderList(){
+    public void showOrderList() {
         orderTable.setItems(getOrderInfoObservableList());
         orderIdCol.setCellValueFactory(new PropertyValueFactory<>("orderId"));
         customerNameCol.setCellValueFactory(new PropertyValueFactory<>("customerName"));
@@ -91,14 +118,16 @@ public class BranchPageController {
 
     @FXML
     public void handleLogoutBtnOnAction(ActionEvent event) throws IOException {
-        Button b = (Button) event.getSource();
-        Stage stage = (Stage) b.getScene().getWindow();
-        FXMLLoader loader = new FXMLLoader(
-                getClass().getResource("/login.fxml")
-        );
-        stage.setScene(new Scene(loader.load(), 1024, 720));
-        LoginController login = loader.getController();
-        stage.show();
+        if(confirmationAlertBox("Confirm to logout?").getResult().equals(ButtonType.OK)) {
+            Button b = (Button) event.getSource();
+            Stage stage = (Stage) b.getScene().getWindow();
+            FXMLLoader loader = new FXMLLoader(
+                    getClass().getResource("/login.fxml")
+            );
+            stage.setScene(new Scene(loader.load(), 1024, 720));
+            loader.getController();
+            stage.show();
+        }
     }
 
     @FXML
@@ -116,22 +145,23 @@ public class BranchPageController {
         popup.showAndWait();
     }
 
-    private void selectedOrderInfo(OrderInfo orderInfo){
+    private void selectedOrderInfo(OrderInfo orderInfo) {
         selectedOrder = orderInfo;
     }
 
 
     @FXML
     public void handleDeleteBtnOnAction(ActionEvent event) throws IOException {
-        if (selectedOrder!=null){
-            //alert
-            //if result.get == buttonType.ok
-            serviceAPI.delete(selectedOrder.getOrderId());
-            showOrderList();
+        if (selectedOrder != null) {
+            if(warningAlertBox("Sure to delete select order?").equals("Delete")){
+                serviceAPI.delete(selectedOrder.getOrderId());
+                showOrderList();
+            }
         }
     }
+
     //Cleaned order page
-    public void showCleanedOrderList(){
+    public void showCleanedOrderList() {
         orderTable2.setItems(getCleanedOrderInfoObservableList());
         orderIdCol2.setCellValueFactory(new PropertyValueFactory<>("orderId"));
         customerNameCol2.setCellValueFactory(new PropertyValueFactory<>("customerName"));
@@ -152,27 +182,29 @@ public class BranchPageController {
         String customerName = customerNameField.getText();
         String customerPhone = customerPhoneField.getText();
         int quantity = Integer.parseInt(clothQuantityField.getText());
-        String date =  LocalDateTime.now().format(formatter);
+        String date = LocalDateTime.now().format(formatter);
 
-        OrderInfo orderInfo = new OrderInfo();
-        orderInfo.setOrderDate(date);
-        orderInfo.setCustomerName(customerName);
-        orderInfo.setCustomerPhone(customerPhone);
-        orderInfo.setBranchName(accountManager.getCurrentBranch().getName());
+        if(confirmationAlertBox("Confirm to add order?").getResult().equals(ButtonType.OK)) {
+            OrderInfo orderInfo = new OrderInfo();
+            orderInfo.setOrderDate(date);
+            orderInfo.setCustomerName(customerName);
+            orderInfo.setCustomerPhone(customerPhone);
+            orderInfo.setBranchName(accountManager.getCurrentBranch().getName());
 
-        Cloth cloth = new Cloth();
-        cloth.setClothQuantity(quantity);
-        cloth.setStatus("Sending to hq");
+            Cloth cloth = new Cloth();
+            cloth.setClothQuantity(quantity);
+            cloth.setStatus("Sending to hq");
 
-        orderInfo.setCloth(cloth);
-        serviceAPI.create(orderInfo);
+            orderInfo.setCloth(cloth);
+            serviceAPI.create(orderInfo);
 
-        clearAddOrderPage();
-        showOrderList();
-        //show alert and change to main page
+            clearAddOrderPage();
+            showOrderList();
+            informationAlertBox("Add order success");
+        }
     }
 
-    private void clearAddOrderPage(){
+    private void clearAddOrderPage() {
         customerNameField.clear();
         customerPhoneField.clear();
         clothQuantityField.clear();
@@ -180,7 +212,7 @@ public class BranchPageController {
 
     //Closed Order Page-------------------------------------------------------------------------------------------------
 
-    public void showClosedOrderList(){
+    public void showClosedOrderList() {
         closedOrderTable.setItems(getClosedOrderInfoObservableList());
         closedOrderIdCol.setCellValueFactory(new PropertyValueFactory<>("orderId"));
         closedCustomerNameCol.setCellValueFactory(new PropertyValueFactory<>("customerName"));
@@ -205,20 +237,30 @@ public class BranchPageController {
     @FXML
     public void handleChangePasswordBtnOnAction(ActionEvent event) throws IOException {
 
-        if(!currentPasswordField.getText().isEmpty()&&!newPasswordField.getText().isEmpty()&&!confirmPasswordField.getText().isEmpty()) {
-            String currentPasswd = currentPasswordField.getText();
-            String newPasswd = newPasswordField.getText();
-            String confirmNewPasswd = confirmPasswordField.getText();
-            if (accountManager.getCurrentBranch() != null) {
-                if (accountManager.getCurrentBranch().getPassword().equals(currentPasswd)) {
-                    if (newPasswd.equals(confirmNewPasswd)) {
-                        accountManager.getCurrentBranch().setPassword(newPasswd);
-                        serviceAPI.updateUserBranch(accountManager.getCurrentBranch());
-                        clearSettingPage();
-                        //alert change success
+        if (!currentPasswordField.getText().isEmpty() &&
+                !newPasswordField.getText().isEmpty() &&
+                !confirmPasswordField.getText().isEmpty()) {
+            if (confirmationAlertBox("Confirm to change password").getResult().equals(ButtonType.OK)) {
+                String currentPasswd = currentPasswordField.getText();
+                String newPasswd = newPasswordField.getText();
+                String confirmNewPasswd = confirmPasswordField.getText();
+                if (accountManager.getCurrentBranch() != null) {
+                    if (accountManager.getCurrentBranch().getPassword().equals(currentPasswd)) {
+                        if (newPasswd.equals(confirmNewPasswd)) {
+                            accountManager.getCurrentBranch().setPassword(newPasswd);
+                            serviceAPI.updateUserBranch(accountManager.getCurrentBranch());
+                            clearSettingPage();
+                            informationAlertBox("Password changed");
+                        }
+                        else errorAlertBox("New password is not match Confirm new password");
+                    }
+                    else {
+                        errorAlertBox("Your current password is incorrect");
                     }
                 }
             }
+        } else {
+            errorAlertBox("Please fill all blank fields");
         }
     }
 
@@ -227,61 +269,97 @@ public class BranchPageController {
         clearSettingPage();
     }
 
-    private void clearSettingPage(){
+    private void clearSettingPage() {
         currentPasswordField.clear();
         newPasswordField.clear();
         confirmPasswordField.clear();
     }
 
     //get order detail
-    private ObservableList<OrderInfo> getOrderInfoObservableList(){
+    private ObservableList<OrderInfo> getOrderInfoObservableList() {
         List<OrderInfo> allOrderInfo = serviceAPI.getAllOrderInfo();
         List<OrderInfo> notCleaningSuccess = new ArrayList<>();
-        for (OrderInfo i : allOrderInfo){
-            if (i.getCloth().getStatus().equals("sending to hq")){
+        for (OrderInfo i : allOrderInfo) {
+            if (i.getCloth().getStatus().equals("Sending to hq")) {
                 notCleaningSuccess.add(i);
             }
         }
-        ObservableList<OrderInfo> orderInfoObservableList = FXCollections.observableArrayList(notCleaningSuccess);
-        return orderInfoObservableList;
+        return FXCollections.observableArrayList(notCleaningSuccess);
     }
-    private ObservableList<OrderInfo> getCleanedOrderInfoObservableList(){
+
+    private ObservableList<OrderInfo> getCleanedOrderInfoObservableList() {
         List<OrderInfo> allOrderInfo = serviceAPI.getAllOrderInfo();
         List<OrderInfo> cleaned = new ArrayList<>();
-        for (OrderInfo i : allOrderInfo){
+        for (OrderInfo i : allOrderInfo) {
             if (i.getCloth().getStatus().equals("Success") || i.getCloth().getStatus().equals("Damaged") ||
-                    i.getCloth().getStatus().equals("No contact response") || i.getCloth().getStatus().equals("Ready to pickup")){
+                    i.getCloth().getStatus().equals("No contact response") || i.getCloth().getStatus().equals("Ready to pickup")) {
                 cleaned.add(i);
             }
         }
-        ObservableList<OrderInfo> orderInfoObservableList = FXCollections.observableArrayList(cleaned);
-        return orderInfoObservableList;
+        return FXCollections.observableArrayList(cleaned);
     }
-    private ObservableList<OrderInfo> getClosedOrderInfoObservableList(){
+
+    private ObservableList<OrderInfo> getClosedOrderInfoObservableList() {
         List<OrderInfo> allOrderInfo = serviceAPI.getAllOrderInfo();
         List<OrderInfo> cleaningSuccess = new ArrayList<>();
-        for (OrderInfo i : allOrderInfo){
-            if (i.getCloth().getStatus().equals("Closed")){
+        for (OrderInfo i : allOrderInfo) {
+            if (i.getCloth().getStatus().equals("Closed")) {
                 cleaningSuccess.add(i);
             }
         }
         return FXCollections.observableArrayList(cleaningSuccess);
     }
 
-    //getter and setter
-    public AccountManager getAccountManager() {
-        return accountManager;
-    }
-
+    //setter
     public void setAccountManager(AccountManager accountManager) {
         this.accountManager = accountManager;
     }
 
-    public WashingOrderServiceAPI getServiceAPI() {
-        return serviceAPI;
-    }
-
     public void setServiceAPI(WashingOrderServiceAPI serviceAPI) {
         this.serviceAPI = serviceAPI;
+    }
+
+    //alert
+    private Alert confirmationAlertBox(String message) {
+        alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirmation");
+        alert.initStyle(StageStyle.UTILITY);
+        alert.setHeaderText(null);
+        alert.setContentText(message);
+        alert.showAndWait();
+        return alert;
+    }
+    private String warningAlertBox(String message){
+        alert = new Alert(Alert.AlertType.WARNING);
+        alert.setTitle("Warning");
+        alert.initStyle(StageStyle.UTILITY);
+        alert.setHeaderText(null);
+        alert.setContentText(message);
+        ButtonType buttonDelete = new ButtonType("Delete");
+        ButtonType buttonCancel = new ButtonType("Cancel", ButtonBar.ButtonData.CANCEL_CLOSE);
+        alert.getButtonTypes().setAll(buttonDelete,buttonCancel);
+        Optional<ButtonType> result = alert.showAndWait();
+        if(result.get() == buttonDelete){
+            return "Delete";
+        }
+        return "Cancel";
+    }
+
+    private void errorAlertBox(String message) {
+        alert = new Alert(Alert.AlertType.ERROR);
+        alert.initStyle(StageStyle.UTILITY);
+        alert.setTitle("ERROR");
+        alert.setHeaderText(null);
+        alert.setContentText(message);
+        alert.showAndWait();
+    }
+
+    private void informationAlertBox(String message){
+        alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.initStyle(StageStyle.UTILITY);
+        alert.setTitle("INFORMATION");
+        alert.setHeaderText(null);
+        alert.setContentText(message);
+        alert.showAndWait();
     }
 }
